Add callbacks tracking and modify order type in PlaceSimpleOrderApp
In the PlaceSimpleOrderApp class, a callback counter has been introduced along with modification in the order attributes. The callback counter is incremented with each successive call, allowing for easier tracking of the lifecycle of an order. The order action has also been altered from 'buy' to 'sell', and the order type has been changed from market to limit orders, necessitating a limit price.