Refactor constants and file utils in IBKR Tools
This commit refactors the file utils and constants in the IBKR Tools. It includes the renaming of '_constants.py' to 'constants.py' and 'confirmation_tools.py' to 'file_utils.py'. Also, improved code for data confirmation and writing to file methods into more object-oriented approach. This will make the functions simpler and easier to use.